// src/app/models/reporte-estudiante/reporte-estudiante.module.ts
import { NgModule } from '@angular/core';
import { CommonModule } from '@angular/common';



@NgModule({
  declarations: [],
  imports: [
    CommonModule
  ]
})
export class ReporteEstudianteModule { 
  Cedula: string;
  Nombre: string;
  Curso: number;
  Paralelo: string;
  Vacuna1: string;
  Vacuna2: string;
  Vacuna3: string;
}